fix(dispatch): cast cost_usd as numeric, forward spawn_status in PATCH route
Two bugs found during the first successful auto-investigation after
the DNS fix:

1. The issue PATCH SQL used `COALESCE($23, 0)` for cost_usd, which
   made PostgreSQL infer $23 as integer (from the `0` literal). When
   the GH Actions workflow sent a float like `2.33`, Postgres rejected
   it with "invalid input syntax for type integer". Fix: cast both
   sides as `::numeric`.

2. The issue PATCH route handler didn't forward `spawn_status` or
   `spawn_run_id` from the request body to `updateIssue()`, so all
   5 workflow types could never update spawn completion status. The
   fields were accepted by `updateIssue()` but never passed through
   the API layer.

+++ lib/services/issueService.ts
@@ -371,7 +371,7 @@ export async function updateIssue(
         spawn_type = COALESCE($20, spawn_type),
         spawn_attempt_count = COALESCE($21, spawn_attempt_count),
         blocked_reason = COALESCE($22, blocked_reason),
-        cost_usd = COALESCE(cost_usd, 0) + COALESCE($23, 0),
+        cost_usd = COALESCE(cost_usd, 0::numeric) + COALESCE($23::numeric, 0),
         model = COALESCE($24, model),
         num_turns = COALESCE(num_turns, 0) + COALESCE($25, 0),
         run_outcome = COALESCE($26, run_outcome),
